change how adding order works

+++ order-service/src/main/java/com/rob/orderservice/dto/OrderRequest.java
@@ -4,5 +4,5 @@ import jakarta.validation.constraints.Positive;
 import lombok.Builder;
 
 @Builder
-public record OrderRequest(int productId, @Positive int quantity, int customerId) {
+public record OrderRequest(int productId, @Positive int quantity, long customerId) {
 }
